fix(mcp): address all PR #864 review findings (B1 S1-S7 H1 H2)
H2: Extract masternode tools to src/mcp/tools/masternode.rs, rename to
masternode_identity_load / masternode_credits_withdraw.

B1: Fix is_platform_address_string panic on non-ASCII input — use
as_bytes() throughout instead of str byte-slicing; add non-ASCII
regression test.

H1 + S1: Add Deserialize + JsonSchema to model::secret::Secret (feature-
gated to mcp/cli). Change MasternodeIdentityLoadParams private-key fields
from String to Secret so keys are mlock-ed from deserialization, before
any SPV wait.

S2: Call decode_identity_id before ensure_spv_synced in the LOAD tool so
a malformed ProTxHash is rejected immediately rather than after a 10-min
sync.

S3: Add impl Display for KeyMode in masternode_input.rs; replace three
literal "owner"/"transfer" sites with .to_string() so the canonical
string is defined once.

S4: Rewrite TC-MN-050 and TC-MN-051 to call
MasternodeCreditsWithdraw::invoke() through a DashMcpService, exercising
key-mode resolution, the payout-address echo, and the SPV gate.

S5: Fix headless shutdown to also drain the initial context's wallet
backend when the network was switched after startup (Arc::ptr_eq guard).

S6: Restructure start_stdio to always call shutdown_wallet_backend even
when serve() or waiting() returns an error.

S7: Dispatch MigrationTask::FinishUnwire from ensure_spv_synced in
standalone/headless mode where the GUI frame-loop never runs.
