+++ src/mcp/mod.rs
@@ -33,14 +33,33 @@ pub async fn start_stdio() -> Result<(), Box<dyn std::error::Error + Send + Sync
     // Keep a clone so we can access the context after `serve()` moves `service`.
     // `DashMcpService` is cheaply cloneable (all fields are `Arc`-wrapped).
     let service_for_shutdown = service.clone();
-    let server = service.serve(rmcp::transport::stdio()).await?;
-    server.waiting().await?;
+
+    // S6: capture the server result WITHOUT short-circuiting via `?` so that
+    // `shutdown_wallet_backend` is ALWAYS called regardless of whether `serve`
+    // or `waiting` returns an error.  The `?` is deferred to after the shutdown.
+    //
+    // `serve().await` errors with `ServerInitializeError`; `waiting().await`
+    // errors with `JoinError` and succeeds with `QuitReason` — both are
+    // mapped to the function's `Box<dyn Error + Send + Sync>` return type.
+    let result: Result<(), Box<dyn std::error::Error + Send + Sync>> = async {
+        let server = service
+            .serve(rmcp::transport::stdio())
+            .await
+            .map_err(|e| -> Box<dyn std::error::Error + Send + Sync> { Box::new(e) })?;
+        server
+            .waiting()
+            .await
+            .map_err(|e| -> Box<dyn std::error::Error + Send + Sync> { Box::new(e) })
+            .map(|_| ())
+    }
+    .await;
 
     // Quiesce the wallet backend (coordinator threads) before returning.
     // The caller's runtime is still alive at this point; the shutdown must
     // complete here, inside `block_on`, NOT during runtime drop.
     service_for_shutdown.shutdown_wallet_backend().await;
-    Ok(())
+
+    result
 }
 
 /// Start the MCP server over HTTP (embedded in GUI app).
