+++ src/mcp/server.rs
@@ -203,8 +203,9 @@ impl DashMcpService {
             .with_async_tool::<tools::identity::IdentityCreditsTransfer>()
             .with_async_tool::<tools::identity::IdentityCreditsWithdraw>()
             .with_async_tool::<tools::identity::IdentityCreditsToAddress>()
-            .with_async_tool::<tools::identity::IdentityMasternodeLoad>()
-            .with_async_tool::<tools::identity::IdentityMasternodeCreditsWithdraw>()
+            // Masternode / evonode tools
+            .with_async_tool::<tools::masternode::MasternodeIdentityLoad>()
+            .with_async_tool::<tools::masternode::MasternodeCreditsWithdraw>()
             // Shielded tools
             .with_async_tool::<tools::shielded::ShieldedShieldFromCore>()
             .with_async_tool::<tools::shielded::ShieldedShieldFromPlatform>()
